% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_padding.R
\name{add_padding}
\alias{add_padding}
\title{Add padding to string}
\usage{
add_padding(d, length = NULL, after = FALSE, pad = "0")
}
\arguments{
\item{d}{vector of strings or numbers}

\item{length}{final string length}

\item{after}{if padding should be added after as opposed to default before}

\item{pad}{padding string of length 1}
}
\value{
vector or character strings of same length.
}
\description{
Add padding to string
}
\examples{
add_padding(sample(1:200,5))
}
