% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_tools.R
\name{cpr_dob}
\alias{cpr_dob}
\title{Extracting date of birth from CPR}
\usage{
cpr_dob(cpr)
}
\arguments{
\item{cpr}{cpr-numbers as ddmmyy\link{-.}xxxx or ddmmyyxxxx. Also mixed formatting. Vector or data frame column.}
}
\value{
vector of Date elements. Format yyyy-mm-dd
}
\description{
For easy calculation.
}
\examples{
cpr_dob("231045-0637")
fsd<-c("010190-2000", "010115-4000", "0101896000","010189-3000","300450-1030","010150-4021")
cpr_dob(fsd)
}
\keyword{cpr}
