% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprs.R
\docType{data}
\name{cprs}
\alias{cprs}
\title{Data frame of 200 cpr numbers}
\format{
A data frame with 200 rows and 1 variable:
\describe{
\item{cpr}{Mixed format cpr-numbers, characters}
}
}
\usage{
data(cprs)
}
\description{
This is just a repeated sample of 8 synthesized cpr-numbers
for testing purposes.
}
\seealso{
\url{https://da.wikipedia.org/wiki/231045-0637}
}
\keyword{datasets}
