% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2dd.R
\name{ds2dd}
\alias{ds2dd}
\title{Data set to data dictionary function}
\usage{
ds2dd(
  ds,
  record.id = "record_id",
  form.name = "basis",
  field.type = "text",
  field.label = NULL,
  include.column.names = FALSE,
  metadata = stRoke::metadata_names
)
}
\arguments{
\item{ds}{data set}

\item{record.id}{name or column number of id variable, moved to first row of
data dictionary, character of integer. Default is "record_id".}

\item{form.name}{vector of form names, character string, length 1 or length
equal to number of variables. Default is "basis".}

\item{field.type}{vector of field types, character string, length 1 or length
equal to number of variables. Default is "text.}

\item{field.label}{vector of form names, character string, length 1 or length
equal to number of variables. Default is NULL and is then identical to field
names.}

\item{include.column.names}{Flag to give detailed output including new
column names for original data set for upload.}

\item{metadata}{Metadata dataframe. Default is the included
stRoke::metadata_names.}
}
\value{
data.frame or list of data.frame and vector
}
\description{
Data set to data dictionary function
}
\examples{
talos$id <- seq_len(nrow(talos))
ds2dd(talos, record.id="id",include.column.names=FALSE)
}
