% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_filter.R
\name{files_filter}
\alias{files_filter}
\title{Filter files in a folder}
\usage{
files_filter(folder.path, filter.by, full.names = TRUE)
}
\arguments{
\item{folder.path}{character. Path of the folder to be filtered}

\item{filter.by}{character. Filter to be applied on the files}

\item{full.names}{logical. Whether to return full file names or not}
}
\value{
character vector. Filtered files
}
\description{
This function filters files in a folder based on the
provided filter.
}
\examples{
# Gives path to files/folders with "tests" in the name in the 
# working directory
files_filter(getwd(),"tests")
}
