% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_stroke.R
\name{generic_stroke}
\alias{generic_stroke}
\title{Generic stroke study outcome}
\usage{
generic_stroke(df, group, score, strata = NULL, variables = NULL)
}
\arguments{
\item{df}{Data set as data frame}

\item{group}{Variable to group by}

\item{score}{Outcome measure variable}

\item{strata}{Optional variable to stratify by}

\item{variables}{String of variable names to include in adjusted OLR-analysis}
}
\value{
Returns list with three elements
}
\description{
Includes table 1, grotta bars and ordinal logistic regression plot.
Please just use this function for illustration purposes.
To dos: modify grottaBar and include as own function.
}
\examples{
# generic_stroke(df = stRoke::talos, group = "rtreat", score = "mrs_6", 
# variables = c("hypertension","diabetes","civil"))
}
