% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plot.R
\name{gt_plot}
\alias{gt_plot}
\title{Plot a gt or gtsummary table as ggplot object}
\usage{
gt_plot(x, ...)
}
\arguments{
\item{x}{A 'gt' or 'gtsummary' table}

\item{...}{Additional arguments to be passed to gt::gtsave}
}
\value{
gg/ggplot element
}
\description{
Relying on \code{library(magick)} and \code{library(webshot2)}
}
\examples{
#gt_plot(mtcars |> gtsummary::tbl_summary())
}
\references{
Please refer
to \href{https://github.com/MSKCC-Epi-Bio/bstfun}{MSKCC-Epi-Bio/bstfun}
}
