% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata_names}
\alias{metadata_names}
\title{Vector of REDCap metadata headers}
\format{
Vector of length 18 with REDCap metadata headers:
\describe{
\item{metadata_names}{characterstrings}
}
}
\usage{
data(metadata_names)
}
\description{
Vector of REDCap metadata headers
}
\seealso{
project-redcap(dot)org (currently the certificate is broken)
}
\keyword{datasets}
