% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_olr.R
\name{plot_olr}
\alias{plot_olr}
\title{Forest plot from ordinal logistic regression.}
\usage{
plot_olr(
  x,
  title = NULL,
  dec = 3,
  lbls = NULL,
  hori = "OR (95 \% CI)",
  vert = "Variables",
  short = FALSE,
  input = c("model", "df")
)
}
\arguments{
\item{x}{input data.}

\item{title}{plot title}

\item{dec}{decimals for labels}

\item{lbls}{labels for variable names. Careful, as the right order is not checked automatically!}

\item{hori}{labels the horizontal axis (this i the y axis as the plot is rotated)}

\item{vert}{labels the horizontal axis (this i the x axis as the plot is rotated)}

\item{short}{flag to half number of ticks on horizontal axis.}

\item{input}{can be either "model", which is a olr model (polr()), or "df", which is a data frame with three columns for OR, lower CI and upper CI.}
}
\value{
gg object
}
\description{
Heavily inspired by https://www.r-bloggers.com/plotting-odds-ratios-aka-a-forrestplot-with-ggplot2/
}
\examples{
iris$ord<-factor(sample(1:3,size=nrow(iris),replace=TRUE),ordered=TRUE)
lm <- MASS::polr(ord~., data=iris, Hess=TRUE, method="logistic")
plot_olr(lm, input="model")
}
\keyword{forest}
\keyword{plot}
