% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_prob.R
\name{print.win_Prob}
\alias{print.win_Prob}
\title{Prints win_prob results}
\usage{
\method{print}{win_Prob}(x, ...)
}
\arguments{
\item{x}{win_prob results.}

\item{...}{ignored for now}
}
\value{
Prints win_prob statistics.
}
\description{
Prints win_prob results
}
