% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stRoke-package.R
\docType{package}
\name{stRoke-package}
\alias{stRoke}
\alias{stRoke-package}
\title{stRoke: Clinical Stroke Research}
\description{
This is an R-toolbox of custom functions for convenient data management and analysis in clinical health research and teaching. The package is mainly collected for personal use, but any use beyond that is encouraged. This package has migrated functions from 'agdamsbo/daDoctoR', and new functions has been added. Version follows months and year. See NEWS/Changelog for release notes. This package includes sampled data from the TALOS trial (Kraglund et al (2018) \doi{10.1161/STROKEAHA.117.020067}). The win_prob() function is based on work by Zou et al (2022) \doi{10.1161/STROKEAHA.121.037744}. The age_calc() function is based on work by Becker (2020) \doi{10.18637/jss.v093.i02}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agdamsbo.github.io/stRoke/}
  \item \url{https://github.com/agdamsbo/stRoke}
  \item Report bugs at \url{https://github.com/agdamsbo/stRoke/issues}
}

}
\author{
\strong{Maintainer}: Andreas Gammelgaard Damsbo \email{agdamsbo@clin.au.dk} (\href{https://orcid.org/0000-0002-7559-1154}{ORCID})

}
\keyword{internal}
