% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talos.R
\docType{data}
\name{talos}
\alias{talos}
\title{Data frame with sample of TALOS data}
\format{
A data frame with 200 rows and 6 variables:
\describe{
\item{rtreat}{Randomisation}
\item{mrs_1}{Modified Rankin scale score at follow-up}
\item{mrs_6}{Modified Rankin scale score at end of study}
\item{hypertension}{Known hypertension}
\item{diabetes}{Known diabetes}
\item{civil}{Cohabitation status}
}
}
\source{
\doi{10.1161/STROKEAHA.117.020067}
}
\usage{
data(talos)
}
\description{
Contains of non-identifiable subset of data from the TALOS trial.
}
\keyword{datasets}
