% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calc.R
\name{age_calc}
\alias{age_calc}
\title{Calculate age in years, months, or days}
\usage{
age_calc(dob, enddate = Sys.Date(), units = "years", precise = TRUE)
}
\arguments{
\item{dob}{Date of birth}

\item{enddate}{End date for age calculation (default is Sys.Date())}

\item{units}{Units for age calculation (default is "years").
Can be c("days", "months", "years")}

\item{precise}{Option to calculate age precisely (default is TRUE)}
}
\value{
Age
}
\description{
Calculate age in years, months, or days
}
\examples{
age_calc(dob = as.Date("1995-04-25"), enddate = as.Date("2020-04-25"), units = "years")
# Kim Larsen (cpr is known from album)
  dob<-as.Date("1945-10-23")
  date<-as.Date("2018-09-30")
  trunc(age_calc(dob,date))
}
\references{
Becker, J.P. (2020). eeptools: An R Package for Teaching and Learning Ecology and Evolutionary Biology. Journal of Statistical Software, 93(2), 1-27. doi: 10.18637/jss.v093.i02
}
\seealso{
\code{\link[lubridate]{time_length}}
}
\keyword{age}
\keyword{date}
\keyword{time}
