% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_tools.R
\name{cpr_check}
\alias{cpr_check}
\title{CPR check}
\usage{
cpr_check(cpr)
}
\arguments{
\item{cpr}{cpr-numbers as ddmmyy"-."xxxx or ddmmyyxxxx.
Also mixed formatting. Vector or data frame column.}
}
\value{
Logical vector of cpr validity
}
\description{
Checking validity of cpr number. Vectorised.
}
\examples{
fsd<-c("2310450637", "010190-2000", "010115-4000",
"300450-1030","010150-4021")
cpr_check("2310450637")
cpr_check(fsd)
all(cpr_check(fsd))
}
\keyword{cpr}
