% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_tools.R
\name{cpr_female}
\alias{cpr_female}
\title{Determine female sex from CPR}
\usage{
cpr_female(cpr)
}
\arguments{
\item{cpr}{cpr-numbers as ddmmyy\link{-.}xxxx or ddmmyyxxxx. Also mixed formatting. Vector or data frame column.}
}
\value{
Logical vector
}
\description{
Just checking if last number of a string is equal or not.
}
\examples{
cpr_female("231045-0637")
}
\keyword{cpr}
