% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_lines.R
\name{source_lines}
\alias{source_lines}
\title{Source only specific lines from an R file}
\usage{
source_lines(file, lines, ...)
}
\arguments{
\item{file}{character string with path to the source file}

\item{lines}{numeric vector of the lines to source from 'file'}

\item{...}{further arguments to be passed to source}
}
\value{
list
}
\description{
Copied from a gist by christophergandrud (https://gist.github.com/christophergandrud/1eb4e095974204b12af9)
This function is really helpfull when working on eg sst.dk-servers. You can not load packages, so you have to re-write this function. It is short, and you will manage!
}
