% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stRoke-package.R
\docType{package}
\name{stRoke-package}
\alias{stRoke}
\alias{stRoke-package}
\title{stRoke: Providing tools for work in clinical stroke research.}
\description{
For a start this package migrates functions from the agdamsbo/daDoctoR-package.
}
\author{
\strong{Maintainer}: Andreas Gammelgaard Damsbo \email{agdamsbo@clin.au.dk} (\href{https://orcid.org/0000-0002-7559-1154}{ORCID})

}
\keyword{internal}
